refactor: move out addConfirmedRequestsToEvents
