+++ ewm/src/main/java/ru/practicum/ewm/compilation/CompilationMapper.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Component;
 import ru.practicum.ewm.compilation.dto.CompilationDto;
 import ru.practicum.ewm.compilation.dto.NewCompilationDto;
 import ru.practicum.ewm.compilation.model.Compilation;
-import ru.practicum.ewm.event.EventMapper;
+import ru.practicum.ewm.event.common.EventMapper;
 import ru.practicum.ewm.event.model.Event;
 
 import java.util.Set;
