+++ ewm/src/main/java/ru/practicum/ewm/event/service/EventPublicService.java
@@ -1,10 +1,7 @@
 package ru.practicum.ewm.event.service;
 
-import com.querydsl.jpa.impl.JPAQuery;
 import ru.practicum.ewm.event.dto.EventFullDto;
 import ru.practicum.ewm.event.dto.EventShortDto;
-import ru.practicum.ewm.event.model.Event;
-import ru.practicum.ewm.event.model.QEvent;
 
 import java.util.List;
 
@@ -25,6 +22,4 @@ public interface EventPublicService {
     );
 
     EventFullDto findFullEventInfo(Long eventId, String clientIp, String endpointPath);
-
-    JPAQuery<Event> setDatesForQuery(String rangeStart, String rangeEnd, JPAQuery<Event> query, QEvent qEvent);
 }
